Fix #8869: Add innerHTML lint check (#8870)
* Add a lint check for innerHTML

* fix _check_for_mandatory_pattern_in_file issue

* fix lint issues

* fix errors from before linter was introduced

* fix path issue

* fix conflicts and merge
