+++ extensions/interactions/NumberWithUnits/protractor.js
@@ -17,8 +17,6 @@
  * Interaction.
  */
 
-var objects = require('../../objects/protractor.js');
-
 var customizeInteraction = function() {
   // There are no customizations.
 };
