+++ extensions/rich_text_components/Image/protractor.js
@@ -17,7 +17,7 @@
  * testing with Protractor.js
  */
 
-var objects = require('../../objects/protractor.js');
+var objects = require(process.cwd() + '/extensions/objects/protractor.js');
 
 var customizeComponent = function(modal, filepath, name) {
   modal.element(by.tagName('button')).click();
